refusal: a rail's structured "no" is the device's vocabulary, not one rail's type
`EncodeStatus` and `ComputeStatus` are the neutral vocabularies for how a
render encode and a compute dispatch ended, and each carried
`MetalBackend(backend::metal::error::Status)` behind
`cfg(all(feature = "backend-metal", target_os = "macos"))`. Two costs: both
enums had two shapes across a feature boundary, and the other rail had no way to
refuse with structure at all — the Vulkan rail's refusals are `&'static str`
slugs, and adopting the richer form would have meant naming Metal's type.

`Status` was never Metal's. It is a `Copy`, fixed-size record of a class, the
registered slug of the check that refused, and up to six load-bearing facts —
`FieldValue` over `u64` / `i64` / `&'static str`. Nothing in it is a Metal
concept. It moves to `backend/refusal.rs` as `RailRefusal`, taking the four
device status codes with it (`backend::metal::abi` defined them and did not use
them; no translation unit in this tree reads them).

The variants become `RailRefused(RailRefusal)`, ungated. `backend::metal::error`
keeps `write_err` and aliases `Status = RailRefusal`, so all ~210 Metal
construction sites are untouched — the alias is what says "this rail's name for
the device's refusal", which is what it always was.

Log vocabulary is unchanged: `class=metal_args` / `class=metal_execute` and
`recovery=metal_failed` still render exactly as before, and the comment beside
them now says they are the refusal's class rather than the rail's, so a
longitudinal grep keeps working.

With this, **the neutral runtime and device model name a rail nowhere outside
`cfg(test)`**. The nine remaining mentions are all in test modules that name a
rail because the test is about one, which `AGENTS.md` requires.

The Vulkan arm gains 3 tests (lib 2197 → 2200): the refusal's own tests now
compile on a build with no Metal rail, which is the point.

Verified: feature matrix all arms, 2767 passed / 0 failed, clippy unchanged on
all three arms.
